+++ types.go
@@ -8,6 +8,7 @@ import (
 
 type secretManagerConfig struct {
 	vault
+	aws
 	// explicitSecrets bool // only get secrets that match the prefix `secret:`
 }